refactor: create separate directive definitions
